--- src/routes/_index/peeps.js
@@ -1,124 +1,0 @@
-export const locations = {
-  berlin: [52.504043, 13.393236],
-  grandRap: [42.9632, -85.6679],
-  bratsk: [56.16667, 101.61667],
-  moscow: [55.75583, 37.61778],
-  ljubljana: [46.05, 14.51667],
-  vegas: [36.1699, -115.1398],
-  prague: [50.08861, 14.42139],
-  seva: [44.6054, 33.5221],
-  hochi: [10.8231, 106.6297],
-  hanover: [52.375, 9.7348],
-  travel: [5.9788, 116.0753],
-  hyderabad: [17.448157, 78.348948],
-  uyo: [7.93, 5.03],
-  kyiv: [50.4016, 30.2525],
-  minsk: [53.8845, 27.4532],
-};
-
-export const peeps = [
-  {
-    name: 'Alexander Lunyov',
-    avatar: 'https://avatars3.githubusercontent.com/u/1477583',
-    url: 'https://github.com/sunify',
-    location: locations.bratsk,
-  },
-  // {
-  //   name: 'Helge Wieding',
-  //   avatar: 'https://avatars3.githubusercontent.com/u/754426',
-  //   url: 'https://github.com/HelgeWieding'
-  // },
-  {
-    name: 'Jan Kremser',
-    avatar: 'https://avatars2.githubusercontent.com/u/41610198',
-    url: 'https://github.com/eezcjkr',
-    location: locations.ljubljana,
-  },
-  {
-    name: 'Evgeni Shavkunov',
-    avatar: 'https://avatars3.githubusercontent.com/u/28968492',
-    url: 'https://github.com/eshavkun',
-    location: locations.prague,
-  },
-  {
-    name: 'Kosta Korenkov',
-    avatar: 'https://avatars2.githubusercontent.com/u/163447',
-    url: 'https://github.com/troggy',
-    location: locations.moscow,
-  },
-  {
-    name: 'Johan Barbie',
-    avatar: 'https://avatars1.githubusercontent.com/u/659301',
-    url: 'https://github.com/johannbarbie',
-    location: locations.berlin,
-  },
-  {
-    name: 'Pinkiebell',
-    avatar: 'https://avatars2.githubusercontent.com/u/40266861',
-    url: 'https://github.com/pinkiebell',
-    location: locations.hanover,
-  },
-  {
-    name: 'Vu Ngoc Quang',
-    avatar: 'https://avatars2.githubusercontent.com/u/24425403',
-    url: 'https://github.com/peara',
-    location: locations.hochi,
-  },
-  // {
-  //   name: 'Steven Pearce',
-  //   avatar: 'https://avatars0.githubusercontent.com/u/36509712',
-  //   url: 'https://github.com/StevenJNPearce'
-  // },
-  // {
-  //   name: 'Sounak Pradhan',
-  //   avatar: 'https://avatars1.githubusercontent.com/u/22390515',
-  //   url: 'https://github.com/sounak98'
-  // },
-  // {
-  //   name: 'Ola Kohut',
-  //   avatar: 'https://avatars3.githubusercontent.com/u/4106590',
-  //   url: 'https://github.com/kohola71'
-  // },
-  {
-    name: 'nanspro',
-    avatar: 'https://avatars0.githubusercontent.com/u/24641089',
-    url: 'https://github.com/nanspro',
-    location: locations.hyderabad,
-  },
-  {
-    name: 'Onuwa Nnachi Isaac',
-    avatar: 'https://avatars1.githubusercontent.com/u/6551094',
-    url: 'https://github.com/iamonuwa',
-    location: locations.uyo,
-  },
-  {
-    name: 'Artico',
-    avatar: 'https://avatars3.githubusercontent.com/u/1811630',
-    url: 'https://github.com/articobandurini',
-    location: locations.kyiv,
-  },
-  {
-    name: 'Maksim Daunarovich',
-    avatar: 'https://avatars3.githubusercontent.com/u/3136647',
-    url: 'https://github.com/MaxStalker',
-    location: locations.minsk,
-  },
-  {
-    name: 'Max Kuck',
-    avatar: 'https://avatars2.githubusercontent.com/u/45236782',
-    url: 'https://github.com/MaxKuck',
-    location: locations.berlin,
-  },
-  {
-    name: 'Tim Daubenschütz',
-    avatar: 'https://avatars0.githubusercontent.com/u/2758453',
-    url: 'https://github.com/TimDaub',
-    location: locations.berlin,
-  },
-  {
-    name: 'vrde',
-    avatar: 'https://avatars1.githubusercontent.com/u/134680',
-    url: 'https://github.com/vrde',
-    location: locations.berlin,
-  },
-];
