+++ src/routes/_interspace/peeps.js
@@ -1,0 +1,26 @@
+export const locations = {
+  berlin: [52.504043, 13.393236],
+  grandRap: [42.9632, -85.6679],
+  bratsk: [56.16667, 101.61667],
+  moscow: [55.75583, 37.61778],
+  ljubljana: [46.05, 14.51667],
+  vegas: [36.1699, -115.1398],
+  prague: [50.08861, 14.42139],
+  seva: [44.6054, 33.5221],
+  hochi: [10.8231, 106.6297],
+  hanover: [52.375, 9.7348],
+  travel: [5.9788, 116.0753],
+  hyderabad: [17.448157, 78.348948],
+  uyo: [7.93, 5.03],
+  kyiv: [50.4016, 30.2525],
+  minsk: [53.8845, 27.4532],
+};
+
+export const peeps = [
+  {
+    name: "Berlin",
+    avatar: "{siteHost}/img/map/berlin.jpeg",
+    url: "https://space.leapdao.org/Berline",
+    location: locations.berlin,
+  },
+];
