+++ src/routes/interspace.svelte
@@ -1,0 +1,14 @@
+<script>
+  import Interspace from "./_interspace/Interspace";
+  import Logo from "../components/Logo";
+</script>
+<style>
+#amoebe {
+display: none;
+}
+</style>
+<div class="content">
+<header>
+<Interspace />
+</header>
+</div>
